start using angular js
We'll use angular in conjunction with websockets in the admin interface
to live update the list of tickets and the chat with each of the
hackers.

This means that we need to stop using the Jinja template to generate the
list of tickets and instead have angular do the API call to the backend
to get the ticket list. Also, when tickets are added in the backend, we
need to emit some kind of socket.io message containing the ticket id in
the __admin__ room. Angular can then perform an API call with the ticket
ID to get the details of the added ticket and then insert it into the
list of tickets on the fly.